requirements.txt
django
djangorestframework
djoser
django-cors-headers
psycopg2

